pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Konstantin Kiselev <konstantin.kiselev.1998@mail.ru>"]
readme = "README.md"
packages = [
  { include = "page_analyzer" },
]
[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
gunicorn = "^22.0.0"
python-dotenv = "^1.0.1"
flask-wtf = "^1.2.1"
psycopg2-binary = "^2.9.9"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
